status offline updates
